requirements.txt
python-dotenv==1.0.0
PyGithub==2.1.1
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.2
pydantic==2.5.2
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1 